Add condition to check existence of remote_#{column}_url.
This applies when determining whether to fire the background
process.

Sometimes you set the image in carrier wave using the
remote_#{column}_url field if the source of the image is remote.

This does not change the column field and hence #{column}_changed?
is set to false and the job is never queued.