Fix removing files on update
During updating, if the `#remove_<column>` was assigned, then
carrierwave_backgrounder would prevent deletion of that photo.

carrierwave_backgrounder doesn't have the functionality of deleting
files in the background (nor should it, this is a very cheap operation).
But it prevents the removal of the file, as if it's going to do it
in the background. Here we put put that, if the file is scheduled for
removal, delegate to the regular foreground process (by assigning the
`process_<column>_upload` attribute).